move conftest to e2e
